Backend/requirements.txt
absl-py==2.3.1
aiofiles==23.1.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.14.0
h5py==3.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.1.0
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
protobuf==4.25.8
pydantic==1.10.23
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==7.4.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.53
starlette==0.48.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
